% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_get.R
\name{family_get_unions}
\alias{family_get_unions}
\alias{family_get_children}
\title{Get links of family}
\usage{
family_get_unions(data)

family_get_children(data)
}
\arguments{
\item{data}{Family data.}
}
\description{
Get links of family
}
