% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_tree.R
\name{family_graph}
\alias{family_graph}
\title{Create family graph}
\usage{
family_graph(data)
}
\arguments{
\item{data}{Family data.}
}
\description{
Create family graph
}
\examples{
library(tidyfamily)
library(tidygraph)

data <- dynasty_capetian()

family_graph(data)
}
